scanner: app.tsx refactor
