--- packages/scanner/src/app.module.css
@@ -1,40 +1,0 @@
-:root {
-    --fs: calc(4vw + 12px);
-    overflow: hidden;
-    overscroll-behavior: none;
-}
-
-button {
-    font-size: var(--fs);
-}
-
-.app {
-    font-size: var(--fs);
-
-    .screen {
-        position: relative;
-        width: 100vw;
-        height: 100vh;
-        max-height: fill-available;
-    }
-}
-
-.auth {
-    display: flex;
-    flex-direction: column;
-    justify-content: space-around;
-    align-items: center;
-    height: 400px;
-    font-size: 2em;
-
-    /*input {*/
-    /*    width: 90vw;*/
-    /*    max-width: 8ch;*/
-    /*    font-size: 10vw;*/
-    /*}*/
-
-    /*button {*/
-    /*    width: 90vw;*/
-    /*    padding: 1em;*/
-    /*}*/
-}
