+++ packages/scanner/src/app/app.module.css
@@ -1,0 +1,5 @@
+:root {
+    --fs: calc(4vw + 12px);
+    overflow: hidden;
+    overscroll-behavior: none;
+}
