+++ packages/scanner/src/components/mock-trans/mock-trans.tsx
@@ -2,13 +2,13 @@ import React, { useContext, useState } from 'react';
 
 import { createMockData } from '~/shared/lib/mock';
 import { db } from '~/db';
-import { AppContext } from '~/app-context';
+import { useApp } from '~/model/app-provider';
 
 import style from './mock-trans.module.css';
 
 export const MockTrans: React.FC = () => {
     const [count, setCount] = useState<number>(1);
-    const { mealTime } = useContext(AppContext);
+    const { mealTime } = useApp();
     const addTestTrans = (count: number, type: 'now' | 'rnd') => {
         if (mealTime) {
             const trans = createMockData(count, type, mealTime);
