+++ packages/scanner/src/components/post-scan/post-scan-anon/post-scan-anon.tsx
@@ -1,15 +1,16 @@
 import type { FC } from 'react';
-import { useContext, useEffect } from 'react';
+import { useEffect } from 'react';
 
-import { AppColor, AppContext } from '~/app-context';
+import { AppColor } from '~/model/app-provider/app-provider';
 import { GreenAnonCard } from '~/components/misc';
+import { useApp } from '~/model/app-provider';
 
 import { useFeedVol } from '../post-scan.utils';
 
 export const PostScanAnon: FC<{
     closeFeed: () => void;
 }> = ({ closeFeed }) => {
-    const { kitchenId, mealTime, setColor } = useContext(AppContext);
+    const { kitchenId, mealTime, setColor } = useApp();
 
     const [doFeed] = useFeedVol(undefined, mealTime, closeFeed, kitchenId);
