+++ packages/scanner/src/index.tsx
@@ -4,7 +4,7 @@ import ReactDOM from 'react-dom/client';
 import './index.css';
 
 import * as serviceWorkerRegistration from './serviceWorkerRegistration';
-import App from './app';
+import App from './app/app';
 import reportWebVitals from './reportWebVitals';
 
 /*mport 'react-hot-loader';
