+++ packages/scanner/src/model/app-provider/index.ts
@@ -1,0 +1,1 @@
+export * from './app-provider'
