+++ packages/scanner/src/model/view-provider/index.ts
@@ -1,0 +1,1 @@
+export * from './view-provider';
