+++ packages/scanner/src/request/use-get-trans.ts
@@ -1,14 +1,14 @@
-import { useCallback, useContext, useMemo, useState } from 'react';
+import { useCallback, useMemo, useState } from 'react';
 import axios from 'axios';
 import dayjs from 'dayjs';
 
 import type { ApiHook } from '~/request/lib';
 import type { ServerTransaction } from '~/db';
 import { db } from '~/db';
-import { AppContext } from '~/app-context';
+import { useApp } from '~/model/app-provider';
 
 export const useGetTrans = (baseUrl: string, pin: string | null, setAuth: (auth: boolean) => void): ApiHook => {
-    const { kitchenId } = useContext(AppContext);
+    const { kitchenId } = useApp();
 
     const [error, setError] = useState<any>(null);
     const [updated, setUpdated] = useState<any>(null);
