+++ packages/scanner/src/screens/stats-screen.tsx
@@ -2,7 +2,7 @@ import cn from 'classnames';
 
 import { ScreenHeader } from '~/components/screen-header';
 
-import css from '../app.module.css';
+import css from '../app/app.module.css';
 import { Stats } from '../components/stats';
 
 export const StatsScreen = () => {
