--- packages/scanner/src/view-context.tsx
@@ -1,11 +1,0 @@
-import React from 'react';
-
-export interface IViewContext {
-    currentView: number;
-    setCurrentView: (any) => void;
-}
-
-// @ts-ignore
-const ViewContext = React.createContext<IViewContext>(null);
-
-export { ViewContext };
